feat: polish — KDoc, README, code review and security fixes
- Add KDoc to all public types (ApiConstants was the only gap)
- Add README.md with usage examples, API docs, and error handling guide
- Replace encodeDefaults with coerceInputValues for safer serialization
- Add UNKNOWN fallback to all enums for forward-compatibility
- Fix reps/distanceMeters/durationSeconds type mismatch (Double→Int per OpenAPI spec)
- Make HevyClientConfig.httpClient internal to prevent resource leakage
- Use expression body and remove redundant @SerialName in UsersApi
- Add ISO 8601 timestamp validation for since/startDate/endDate params
- Install SDK plugins on caller-provided HttpClient via .config{}
- Sanitize server error strings (strip control chars, truncate to 500)
- Document coerceInputValues trade-off in SdkJson KDoc

+++ src/commonMain/kotlin/com/hevy/sdk/model/routine/RoutineSet.kt
@@ -10,10 +10,10 @@ data class RoutineSet(
     val index: Int,
     val type: SetType,
     @SerialName("weight_kg") val weightKg: Double? = null,
-    val reps: Double? = null,
+    val reps: Int? = null,
     @SerialName("rep_range") val repRange: RepRange? = null,
-    @SerialName("distance_meters") val distanceMeters: Double? = null,
-    @SerialName("duration_seconds") val durationSeconds: Double? = null,
+    @SerialName("distance_meters") val distanceMeters: Int? = null,
+    @SerialName("duration_seconds") val durationSeconds: Int? = null,
     val rpe: Double? = null,
     @SerialName("custom_metric") val customMetric: Double? = null,
 )
