+++ src/commonMain/kotlin/com/hevy/sdk/model/workout/WorkoutSet.kt
@@ -10,9 +10,9 @@ data class WorkoutSet(
     val index: Int,
     val type: SetType,
     @SerialName("weight_kg") val weightKg: Double? = null,
-    val reps: Double? = null,
-    @SerialName("distance_meters") val distanceMeters: Double? = null,
-    @SerialName("duration_seconds") val durationSeconds: Double? = null,
+    val reps: Int? = null,
+    @SerialName("distance_meters") val distanceMeters: Int? = null,
+    @SerialName("duration_seconds") val durationSeconds: Int? = null,
     val rpe: Double? = null,
     @SerialName("custom_metric") val customMetric: Double? = null,
 )
